+++ game/test_items.py
@@ -7,11 +7,7 @@ Contributors:
 """
 
 
-from common import Item, Action, Stats, null_uuid
-from lang import get_lang_choice
-
-
-text = get_lang_choice()
+from game_classes import Item, Action, Stats, null_uuid
 
 
 """ Item
